Refactoring in EV3DevPlatforms
